ai-pipeline/lightning_app/requirements.txt
# PhotoGenius AI - Lightning App Requirements

# Lightning
lightning>=2.1.0
lightning-cloud>=0.5.0

# PyTorch & CUDA
torch>=2.0.0
torchvision>=0.15.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# LoRA Training
peft>=0.7.0
bitsandbytes>=0.41.0

# Face Detection & Recognition
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
opencv-python>=4.9.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0

# Memory Optimization
xformers>=0.0.23

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
httpx>=0.25.0
python-multipart>=0.0.6
